requirements.txt
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
recbole>=1.2.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
transformers>=4.35.0
openai>=1.0.0
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.13.0
tqdm>=4.65.0
pyyaml>=6.0
